% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{apio}
\alias{apio}
\title{Apply function with invisible output}
\usage{
apio(x, fun, ...)
}
\arguments{
\item{x}{vector or list}

\item{fun}{function}

\item{...}{parameters passed on to FUN}
}
\description{
Performs a loop over a vector or a list, performing an action on each element.
The output is not displayed in the terminal
}
\details{
The name is short for: s(ap)ply with (i)invisible (o)utput
}

