% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{apwi}
\alias{apwi}
\title{Apply function with an iterator}
\usage{
apwi(X, FUN, ...)
}
\arguments{
\item{FUN}{function. The structure of the function should be function(i, x)
where 'i' is an index variable and 'x' is the data.}

\item{...}{parameters passed on to FUN}

\item{x}{vector or list}
}
\description{
Performs a loop over a vector or a list, performing an action on each element.
The action has access to an indexing variable marking position in x.
}
\details{
To provide the iterator, the implementation actually uses a for loop.

The names is short for: (ap)ply (w)ith (i)terator
}

