% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{assignc}
\alias{assignc}
\title{Assign value to a variable using value from cache}
\usage{
assignc(x, overwrite = FALSE, warn = FALSE)
}
\arguments{
\item{x}{character, name of object to retrieve from cache}

\item{overwrite}{logical, when TRUE, load from cache occurs regardless
of whether object already exists; when FALSE, assign only happens}

\item{warn}{logical, determine if show warning when assignment fails}
}
\value{
integer code, 0 when load failed; 1 if object loaded from cache; 2 if assign
was aborted because object already exists
}
\description{
Assign value to a variable using value from cache
}
