% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cachedir}
\alias{cachedir}
\title{Set or query a cache directory}
\usage{
cachedir(dir = NULL)
}
\arguments{
\item{dir}{character, path to cache directory (will be created if
not exists)}
}
\description{
Set or query a cache directory
}
