% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cachefile}
\alias{cachefile}
\title{Get path to a cache file matching object}
\usage{
cachefile(x)
}
\arguments{
\item{x}{character, object name of interest}
}
\value{
character, path to file
}
\description{
Get path to a cache file matching object
}
