% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheprefix}
\alias{cacheprefix}
\title{Set or query a prefix for cache files}
\usage{
cacheprefix(prefix = NULL)
}
\arguments{
\item{prefix}{character, prefix for all cache files}
}
\description{
Set or query a prefix for cache files
}
