% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty}
\alias{empty}
\title{evaluate if object is empty (length 0)}
\usage{
empty(x)
}
\arguments{
\item{x}{object, vector, list, matrix}
}
\value{
logical value, TRUE if object is of length zero
}
\description{
The name is short for: object is (empty)
}
