% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{existsc}
\alias{existsc}
\title{Check if an object exists in cache}
\usage{
existsc(x)
}
\arguments{
\item{x}{character, name of object to lookup in cache}
}
\value{
logical, TRUE if matchin file exists in cache
}
\description{
Check if an object exists in cache
}
