% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grepf}
\alias{grepf}
\title{Pattern matching inside text files}
\usage{
grepf(pattern, path = ".", file.pattern = NULL, ...)
}
\arguments{
\item{pattern}{character, pattern to look for}

\item{path}{directory to look in}

\item{file.pattern}{pattern to consider among the files}

\item{...}{other parameters passed on to list.files()}
}
\description{
Reads contents of all files in a directory and matches a pattern line-by-line.
}
\details{
The name is short for: (grep) inside (f)iles
}
