% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{grepv}
\alias{grepv}
\title{Pattern matching}
\usage{
grepv(pattern, x, value = T, ...)
}
\arguments{
\item{pattern}{character, pattern to look for}

\item{x}{character vector, object wherein to look for the pattern}

\item{value}{logical, set to T to return the values in that
match the pattern}

\item{...}{parameters passed on to grep}
}
\description{
This is a wrapper for grep. By default it returns the matching
values as opposed to the indeces where the hits occur.
}
\details{
The name is short for: (grep) returning (v)alues.
}

