% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h1}
\alias{h1}
\title{Return first element}
\usage{
h1(x, ...)
}
\arguments{
\item{x}{vector, matrix, data-frame, or nany object that can be
processed by head()}

\item{...}{parameters passed on to head}
}
\description{
The name is short for: (h)ead with n=(1)
}
