% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h2}
\alias{h2}
\title{Return second element}
\usage{
h2(x, ...)
}
\arguments{
\item{x}{vector, matrix, data-frame, or any object that can be
processed by head()}

\item{...}{parameters passed on to head}
}
\description{
The name is short for: (h)ead with n=(2)
}
