% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h3}
\alias{h3}
\title{Return third element}
\usage{
h3(x, ...)
}
\arguments{
\item{x}{vector, matrix, data-frame, or any object that can be
processed by head()}

\item{...}{parameters passed on to head}
}
\description{
The name is short for: (h)ead with n=(3)
}
