% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/projects.R
\name{initp}
\alias{initp}
\title{Initialize a directory structure for a project}
\usage{
initp(workdir, subdirs = c("notes", "figures", "tables", "Rda", "data",
  "code"), verbose = TRUE, ...)
}
\arguments{
\item{workdir}{directory path, root directory for new project}

\item{subdirs}{character vector, names for subdirectories}

\item{verbose}{logical. If TRUE the function outputs progress messages.}

\item{...}{additional arguments passed on to dir.create, e.g. permission
codes. See dir.create for details}
}
\description{
The name is short for: (init)ialize a new (p)roject
}

