% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ji}
\alias{ji}
\title{compute Jaccard Index of two sets}
\usage{
ji(a, b)
}
\arguments{
\item{a}{vector of set elements}

\item{b}{vector of set elements}
}
\description{
The Jaccard Index of two sets is defined as the ratio of the
size of the set intersection (common element) and the
size of the set union (all elements).
}
\details{
The name is short for: (j)accard (i)ndex
}
\examples{
ji(c(1,2,3), c(1,4,5)) # one common element out of five, JI=0.2 

}
