% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lengrep}
\alias{lengrep}
\title{length of grep result}
\usage{
lengrep(pattern, x, ...)
}
\arguments{
\item{pattern}{character string with a grep pattern}

\item{x}{character vector to search in}

\item{...}{any other arguments passed on to grep()}
}
\value{
integer, length of grep result
}
\description{
length of grep result
}
