% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{limq}
\alias{limq}
\title{create an interval to catch a set proportion of input data based on quantiles}
\usage{
limq(
  data,
  quantiles = c(0.01, 0.99),
  lim = c(NA, NA),
  column = NA,
  pretty = c(NA, NA)
)
}
\arguments{
\item{data}{list, data-frame, or matrix with input data}

\item{quantiles}{numeric of length 2, used in quantiles()}

\item{lim}{numeric of length 2, starting interval, values with NA are filled in
and positions with non-NA are left alone.}

\item{column}{identifier for column}

\item{pretty}{numeric of length 2, ensures that limits are multiples of these values}
}
\value{
numeric vector of length 2 with (min, max) values for limits
}
\description{
The name is short for: (lim)its to capture (q)uantiles
}
