% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linechars}
\alias{linechars}
\title{Count characters-per-line in files}
\usage{
linechars(path = ".", file.pattern = NULL, min.chars = 0)
}
\arguments{
\item{path}{character, directory path}

\item{file.pattern}{character, filename match pattern}

\item{min.chars}{interger, minimal number of characters on a line.
Shorter lines will be omitted from the output}
}
\value{
list, each element summarizing line lengths in one file
}
\description{
The name is short for: (line) (char)acters
}
