% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{load1}
\alias{load1}
\title{Load one R object from an Rda file}
\usage{
load1(file)
}
\arguments{
\item{file}{filename with Rda object}
}
\description{
This functions loads an object from an Rda file. In contrast to
standard load(), this function does not put the loaded object
in the current environment. Rather, this function returns the loaded
data as an object. It is thus possible to save an object x using
save(x, file=myfile), and then retrieve it using y = load1(myfile).
}
\details{
The name is short for: (load 1) object
}

