% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{loadc}
\alias{loadc}
\title{Retrieve object from cache}
\usage{
loadc(x, envir = TRUE)
}
\arguments{
\item{x}{character, name of object to retrieve from cache}

\item{envir}{logical, whether to assign cache value to an object
in calling environment}
}
\description{
Look up content in the cache. By default, create or overwrite an
object in calling environment with contents from cache
}
