% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{loaddir}
\alias{loaddir}
\title{Load contents of directory (from multiple files of different types)}
\usage{
loaddir(dirpath, extensions = c("Rda", "txt", "txt.gz", "json", "json.gz",
  "tsv", "tsv.gz", "csv", "csv.gz"), verbose = FALSE, ...)
}
\arguments{
\item{dirpath}{path for input directory}

\item{extensions}{vector of extensions to try to load}

\item{verbose}{logical, prints filenames as it is loading them}

\item{...}{other parameters passed on to rtht}
}
\description{
Scans a directory and loads data from all files,
using an appropriate loader for each file extension type.
}
