% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lst}
\alias{lst}
\title{Collect objects into a named list}
\usage{
lst(...)
}
\arguments{
\item{...}{several arguments}
}
\value{
a named list holding to input objects
}
\description{
The names is short for: (l)i(st) from named entities
}
\examples{
a = 4
b = 3
lst(a, b)

}
