% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msgs.R
\name{msg}
\alias{msg}
\title{Message logging function}
\usage{
msg(x = NULL, ..., action = c("log", "get", "show", "reset"))
}
\arguments{
\item{x}{character, messages to log}

\item{...}{character, this allows the user to write a
message that is split into multiple parts, e.g. msg("hello", "there")}

\item{action}{character, determines what the processing
performed on the input. Defaults to 'log', which logs a new message.
Other options are show and get, which retrieve all the stored message.
Action 'reset' deletes all messages and clears up memory.}
}
\description{
This is an interface to a simple message logging system. The
basic usage is to call msg("text") to output some text onto the
console with cat. This function records all the messages logged, so
that it is possible to retrive them using msg(NULL, action="show")
}

