% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{namesV}
\alias{namesV}
\title{Get element names in a vector that hold given values}
\usage{
namesV(x, v)
}
\arguments{
\item{x}{vector}

\item{v}{vector of value within x to look for}
}
\description{
The name is short for: (names) of elements with given (V)alues
}

