% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{newmsg}
\alias{newmsg}
\title{Create new message logging object}
\usage{
newmsg(verbose = TRUE, date = TRUE)
}
\arguments{
\item{verbose}{logical, determines if message passed to the
logging function are sent back to console via 'message'}

\item{date}{logical, determines if messages are recorded
with current date and time}
}
\description{
Create new message logging object
}
