% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newv}
\alias{newv}
\title{Create new empty object}
\usage{
newv(mode = "list", length = 0, names = NULL)
}
\arguments{
\item{mode}{character, specifies what kind of vector to create}

\item{length}{integer, length of vector to create}

\item{names}{character, vector of names to create}
}
\description{
This is a wrapper for vector() followed by setting names
}
\details{
The name is short for: (new) (v)ector
}

