% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlist}
\alias{nlist}
\title{Create a named list using an existing object or data frame}
\usage{
nlist(df = NULL, values.col = NULL, names.col = NULL)
}
\arguments{
\item{df}{object, can be vector, matrix, or data.frame}

\item{values.col}{character, identifies source of values for the list}

\item{names.col}{charcter, identifies source of names for the list}
}
\value{
named list
}
\description{
The name is short for: (n)amed (list)
}
\examples{
nlist(c("a", "b", "c"))
nlist(values=c(1,2,3), names=c("x", "y", "z"))

}
