% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nvec}
\alias{nvec}
\title{Create a named vector using two columns in a data-frame}
\usage{
nvec(df, values.col = NULL, names.col = NULL)
}
\arguments{
\item{df}{data frame}

\item{values.col}{character, column in df used for vector values}

\item{names.col}{character, column in df used for vector names}
}
\description{
The name is short for: (n)amed (vec)tor
}
\examples{
df = data.frame(A=letters[1:4], B=c(4,2,3,1))
nvec(df, "B", "A")

}
