% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{oi}
\alias{oi}
\title{compute Overlap Index of two sets}
\usage{
oi(a, b)
}
\arguments{
\item{a}{vector of set elements}

\item{b}{vector of set elements}
}
\description{
The Overlap Index of two sets is defined as the ratio of the
size of the set intersection (common elements) and the size
of the primary set.
}
\examples{
oi(c(1,2), c(2,3,4,5)) # one element out of two in "a" is present in "b"

}
