% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluck}
\alias{pluck}
\title{Extract a component from an object}
\usage{
pluck(x, n, type = c("auto", "character", "integer", "numeric",
  "logical", "list", "matrix", "data.frame"))
}
\arguments{
\item{x}{object - either vector, list, matrix, etc}

\item{n}{integer or string that identifies a part of x}

\item{type}{string, identifies}
}
\description{
Extract a component from an object
}
