% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluck1}
\alias{pluck1}
\title{Extract the first component from an object}
\usage{
pluck1(x, type = class(x))
}
\arguments{
\item{x}{object - either vector, list, matrix, etc}

\item{type}{string, see pluck() for details}
}
\description{
Extract the first component from an object
}
