% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluck2}
\alias{pluck2}
\title{Extract the second component from an object}
\usage{
pluck2(x, type = class(x))
}
\arguments{
\item{x}{object - either vector, list, matrix, etc}

\item{type}{string, see pluck() for details}
}
\description{
Extract the second component from an object
}
