% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{rmc}
\alias{rmc}
\title{Remove an object and its disk-cached match}
\usage{
rmc(x)
}
\arguments{
\item{x}{object to remove}
}
\description{
Wrapper for rm(), which in addition attempts to remove a matched
file from disk cache (see cachedir())
}
\details{
WARNING: This function will remove files from the file system;
use with caution.
}
