% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{rtht}
\alias{rtht}
\title{Read table from file}
\usage{
rtht(f, rowid.column = NULL, header = TRUE, sep = "\\t",
  comment.char = "#", stringsAsFactors = FALSE, ...)
}
\arguments{
\item{f}{filename}

\item{rowid.column}{character name of column to use as rownames}

\item{header}{logical, determines if to read the header}

\item{sep}{character, column separator}

\item{comment.char}{character denoting line-to-ignore (comment)}

\item{stringsAsFactors}{logical, determines whether strings are
converted into factors}

\item{...}{other parameters passed on to read.table}
}
\description{
A wrapper for read.table. By default, it assumes the file
has a header row, is tab separated. By default it does not
convert strings into factors.
}
\details{
The name is short for: (r)ead (t)able with (h)eader and (t)ab separated
}
