% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{savec}
\alias{savec}
\title{Save an object into disk-based cache}
\usage{
savec(x)
}
\arguments{
\item{x}{object to save}
}
\value{
full path to the saved object
}
\description{
Wrapper for save(), which write an Rda representation of an
object into a matched file in disk cache (see cache())
}
