% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setsummary}
\alias{setsummary}
\title{Characterize memberships in two sets}
\usage{
setsummary(a, b, unique.force = TRUE)
}
\arguments{
\item{a}{vector (character, numeric, etc.) with members of set a}

\item{b}{vector with members of set b}

\item{unique.force}{logical, set TRUE to enforce that a, b have
non-repeating elements (i.e. they are sets), set FALSE to use
a,b as-is, i.e. potentially with non-unique elements.}
}
\value{
list holding three vectors and a summary. Vectors contain
the set elements that are private and common in a, b, and a vector
with membership counts. Warning, membership counts might not be
reliable if a, b have duplicate entries
}
\description{
Characterize memberships in two sets
}
