% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{widen}
\alias{widen}
\title{Widen an interval by a relative amount}
\usage{
widen(interval, by = c(0.05, 0.05), log = FALSE)
}
\arguments{
\item{interval}{numeric vector of length 2}

\item{by}{numeric vector of length 2,
percentage widening on each side of the interval}

\item{log}{boolean, if TRUE, widen so that "by" percentage
is perceived on a logarithmic scale}
}
\description{
This can be useful while specifying the x/y limits for a plot
}
