% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{x2col}
\alias{x2col}
\title{Convert numbers or matrices into colors with transparency}
\usage{
x2col(x, col = c("#0000ff", "#ff0000"), maxval = 1, bgcol = "#ffffff")
}
\arguments{
\item{x}{numeric vector or matrix.}

\item{col}{vector of two colors in #XXXXXX format.
First element determines color associated with negative values.
Second element determines color associated with positive values}

\item{maxval}{numeric. Value for which color reaches saturation.}

\item{bgcol}{color of background. Colors with full transparency
will be replaced with this bg color instead.}
}
\description{
Convert numbers or matrices into colors with transparency
}

