% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x2df}
\alias{x2df}
\title{convert an object to a data.frame}
\usage{
x2df(x, stringsAsFactors = F)
}
\arguments{
\item{x}{object, could be a matrix, data.table, etc.}

\item{stringsAsFactors}{logical, convert character-columns to factors or not}
}
\description{
The names is short for: convert (x) (to) a (d)ata (f)frame. Useful for
generating transparency levels.
}
