% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msgs.R
\name{newmsg}
\alias{newmsg}
\title{Create new message logging object}
\usage{
newmsg(verbose = TRUE, date = TRUE)
}
\arguments{
\item{verbose}{logical, determines if message passed to the
logging function are 'cat' back to console}

\item{date}{logical, determines if messages are recorded
with current date and time}
}
\description{
Create new message logging object
}

