% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{rtht}
\alias{rtht}
\title{Read table from file}
\usage{
rtht(f, header = TRUE, sep = "\\t", stringsAsFactors = FALSE, ...)
}
\arguments{
\item{f}{filename}

\item{header}{logical, determines if to read the header}

\item{sep}{character, column separator}

\item{stringsAsFactors}{logical, determines whether strings are
converted into factors}

\item{...}{other parameters passed on to read.table}
}
\description{
This is a wrapper for read.table. By default it assumes the file
has a header row, is tab separated. By default it does not
convert strings into factors.
}
\details{
The name is short for: (r)ead (t)able with (h)eader and (t)ab separated
}

