% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{x2hex}
\alias{x2hex}
\title{Convert numbers [0,1] into two-digit hex [00,ff]}
\usage{
x2hex(x, type = "x")
}
\arguments{
\item{x}{numeric vector}

\item{type}{use "x" for lowercase hex and "X" for uppercase hex}
}
\description{
The names is short for: convert (x) (to) a (hex) value. Useful for
generating transparency levels.
}

